# docs/swagger.yaml
basePath: /api/v1
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
  github_com_mr-kaynak_go-core_internal_core_errors.ErrorCode:
    enum:
    - AUTH_001
    - AUTH_002
    - AUTH_003
    - AUTH_004
    - AUTH_005
    - AUTH_006
    - AUTH_007
    - USER_001
    - USER_002
    - USER_003
    - USER_004
    - VAL_001
    - VAL_002
    - VAL_003
    - VAL_004
    - DB_001
    - DB_002
    - DB_003
    - DB_004
    - DB_005
    - EXT_001
    - EXT_002
    - EXT_003
    - EXT_004
    - EMAIL_001
    - EMAIL_002
    - MQ_001
    - MQ_002
    - INTERNAL_001
    - NOT_FOUND
    - METHOD_NOT_ALLOWED
    - CONFLICT
    - BAD_REQUEST
    - UNPROCESSABLE_ENTITY
    type: string
    x-enum-varnames:
    - CodeUnauthorized
    - CodeInvalidCredentials
    - CodeTokenExpired
    - CodeTokenInvalid
    - CodeInsufficientRights
    - CodeAccountLocked
    - CodeAccountNotActivated
    - CodeUserNotFound
    - CodeUserAlreadyExists
    - CodeInvalidUserData
    - CodeEmailNotVerified
    - CodeValidationFailed
    - CodeInvalidInput
    - CodeMissingField
    - CodeInvalidFormat
    - CodeDatabaseError
    - CodeRecordNotFound
    - CodeDuplicateEntry
    - CodeConstraintViolated
    - CodeTransactionFailed
    - CodeServiceUnavailable
    - CodeTimeout
    - CodeRateLimitExceeded
    - CodeCircuitBreakerOpen
    - CodeEmailSendFailed
    - CodeInvalidEmail
    - CodeMessagePublishFailed
    - CodeMessageConsumeFailed
    - CodeInternal
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeConflict
    - CodeBadRequest
    - CodeUnprocessableEntity
  github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail:
    properties:
      code:
        $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ErrorCode'
      detail:
        type: string
      instance:
        type: string
      meta:
        additionalProperties: true
        type: object
      status:
        type: integer
      time:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
...
  internal_modules_identity_api.BulkPolicyRequest:
    properties:
      policies:
        items:
          $ref: '#/definitions/internal_modules_identity_api.PolicyRequest'
        minItems: 1
        type: array
    required:
    - policies
    type: object
  internal_modules_identity_api.CheckPermissionRequest:
    properties:
      action:
        type: string
      domain:
        type: string
      object:
        type: string
      subject:
        type: string
    required:
    - action
    - object
    - subject
    type: object
...
  internal_modules_identity_api.PolicyRequest:
    properties:
      action:
        type: string
      domain:
        type: string
      effect:
        enum:
        - allow
        - deny
        type: string
      object:
        type: string
      subject:
        type: string
    required:
    - action
    - domain
    - effect
    - object
    - subject
    type: object
...
  internal_modules_identity_api.ResourceGroupRequest:
    properties:
      domain:
        type: string
      group:
        type: string
      resource:
        type: string
    required:
    - group
    - resource
    type: object
...
  internal_modules_identity_api.UserRoleRequest:
    properties:
      domain:
        type: string
      role:
        type: string
    required:
    - role
    type: object
...
host: localhost:3000
info:
  contact:
    name: Go-Core Team
    url: https://github.com/mr-kaynak/go-core
  description: Enterprise-grade Go boilerplate with REST API, gRPC services, and notification
    system
  license:
    name: MIT
  title: Go-Core API
  version: 1.0.0
paths:
...
  /policies:
    delete:
      consumes:
      - application/json
      description: Remove an existing Casbin policy rule
      parameters:
      - description: Policy to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.PolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Policy removed
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Remove a policy
      tags:
      - Policies
    post:
      consumes:
      - application/json
      description: Add a new Casbin policy rule
      parameters:
      - description: Policy definition
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.PolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Policy added
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Add a policy
      tags:
      - Policies
  /policies/bulk:
    post:
      consumes:
      - application/json
      description: Add multiple policies at once
      parameters:
      - description: Bulk policy request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.BulkPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bulk operation result
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Bulk add policies
      tags:
      - Policies
  /policies/check:
    post:
      consumes:
      - application/json
      description: Check if a subject has permission to perform an action
      parameters:
      - description: Permission check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.CheckPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission check result
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Check permission
      tags:
      - Policies
  /policies/reload:
    get:
      description: Reload all policies from database
      produces:
      - application/json
      responses:
        "200":
          description: Policies reloaded
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Reload policies
      tags:
      - Policies
  /policies/resource-groups:
    delete:
      consumes:
      - application/json
      description: Remove a resource from a resource group
      parameters:
      - description: Resource group to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.ResourceGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Resource removed from group
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Remove resource from group
      tags:
      - Policies
    post:
      consumes:
      - application/json
      description: Add a resource to a resource group
      parameters:
      - description: Resource group assignment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.ResourceGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Resource added to group
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Add resource to group
      tags:
      - Policies
  /policies/roles/{role}/users:
    get:
      description: Get all users with a specific role
      parameters:
      - description: Role name
        in: path
        name: role
        required: true
        type: string
      - default: default
        description: Domain filter
        in: query
        name: domain
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users with role
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Get users for role
      tags:
      - Policies
  /policies/save:
    post:
      description: Save current policies to database
      produces:
      - application/json
      responses:
        "200":
          description: Policies saved
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Save policies
      tags:
      - Policies
  /policies/users/{user_id}/permissions:
    get:
      description: Get all permissions for a user in a domain
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - default: default
        description: Domain filter
        in: query
        name: domain
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User permissions
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Get user permissions
      tags:
      - Policies
  /policies/users/{user_id}/roles:
    delete:
      consumes:
      - application/json
      description: Remove a role from a user in a domain
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: Role to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.UserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role removed
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Remove role from user
      tags:
      - Policies
    get:
      description: Get all roles assigned to a user
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - default: default
        description: Domain filter
        in: query
        name: domain
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User roles
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Get user roles
      tags:
      - Policies
    post:
      consumes:
      - application/json
      description: Assign a role to a user in a domain
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      - description: Role assignment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.UserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role added
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Add role to user
      tags:
      - Policies
...
securityDefinitions:
  Bearer:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"